refac: rename layouts folder to components
