+++ src/components/Friend.js
